remove blackbox
