+++ hlib/python/hlib/__init__.py
@@ -1,3 +1,4 @@
 from . import op
+from . import ip
 from . import frontend
 from . import utils
